+++ ntex/src/ws/client.rs
@@ -12,7 +12,7 @@ use crate::connect::rustls::{TlsClientFilter, TlsConnector};
 use tls_rustls::ClientConfig as RustlsClientConfig;
 
 use base64::{Engine, engine::general_purpose::STANDARD as base64};
-use nanorand::{Rng, WyRand};
+use nanorand::Rng;
 
 use crate::client::{ClientCodec, ClientConfig, ClientRawRequest, ClientResponse};
 use crate::connect::{Connect, ConnectError, Connector};
@@ -203,7 +203,7 @@ where
         // a base64-encoded (see Section 4 of [RFC4648]) value that,
         // when decoded, is 16 bytes in length (RFC 6455)
         let mut sec_key: [u8; 16] = [0; 16];
-        WyRand::new().fill(&mut sec_key);
+        nanorand::tls_rng().fill(&mut sec_key);
         let key = base64.encode(sec_key);
 
         head.headers.insert(